+++ lib/fft_burst_tagger_impl.h
@@ -73,7 +73,7 @@ namespace gr {
       float * d_burst_mask_f;
       float * d_ones_f;
       float d_threshold;
-      float d_center_frequency;
+      double d_center_frequency;
       float d_window_enbw;
 
       FILE * d_burst_debug_file;
@@ -98,7 +98,7 @@ namespace gr {
       void tag_gone_bursts(int noutput_items);
 
      public:
-      fft_burst_tagger_impl(float center_frequency, int fft_size, int sample_rate,
+      fft_burst_tagger_impl(double center_frequency, int fft_size, int sample_rate,
                             int burst_pre_len, int burst_post_len, int burst_width,
                             int max_bursts, float threshold, int history_size,
                             bool offline, bool debug);
